feat: add hook to secondary vacancy component and rename folder
