+++ job-search/src/components/SecondaryVacancy/SecondaryVacancy.module.scss
