+++ job-search/src/components/SecondaryVacancy/index.ts
